requirements.txt
django<5.0,>=4.0
celery<6.0,>=5.2
django-celery-results<3.0,>=2.4
redis<5.0,>=4.0
requests<3.0,>=2.26
djangorestframework<4.0,>=3.12
psycopg2-binary<3.0,>=2.9
